refactored as per pull request
