+++ exercises/01-elixir/02-data-types/03-lists/01-creation/solution2.exs
@@ -1,0 +1,5 @@
+defmodule Util do
+  def range(a, b), do: calcrange(a, b, [])
+  defp calcrange(a, b, acc) when b < a, do: acc
+  defp calcrange(a, b, acc), do: calcrange(a, b - 1, [b | acc])
+end
